chore: change PlutoConnectionProps location
